- Moved frequently used functions from main.py to custom_functions.py - Changed imports accordingly - Removed unused imports
