+++ main.py
@@ -6,8 +6,6 @@ import asyncio
 
 import json
 import mysql.connector
-from pytz import timezone
-from datetime import datetime
 
 import discord
 from discord import Embed, Color, Intents, Activity, ApplicationContext, Guild
@@ -16,7 +14,8 @@ from discord.errors import Forbidden
 from discord.ext.commands import errors
 
 from custom_errors import NotEnoughMoney, UserIsBot, UnknownObject, MaxAmountReached, IncorrectBetValue, \
-    InvalidTimeString, HowDidYouGetHere, CommandDisabled, CantBuyTurnip, CantSellTurnip
+    InvalidTimeString, HowDidYouGetHere, CommandDisabled, CantBuyTurnip, CantSellTurnip, UnknownCode, CodeLimitReached
+from custom_functions import time_now, get_text, get_parameter, var_set
 
 
 load_dotenv()
@@ -52,168 +51,6 @@ async def new_player(user):
         curC.execute(f"INSERT INTO stats (discordID) VALUES ({user.id})")
         curD.execute(f"INSERT INTO potato (discordID) VALUES ({user.id})")
         db.commit()
-
-
-def time_now(tz: str = "Europe/Paris"):
-    """
-        Return current time
-    """
-    return datetime.now(tz=timezone(tz))
-
-
-def get_parameter(param):
-    """
-        Get config parameters
-    """
-    with open("json/config.json", "r", encoding="utf-8") as config_file:
-        config = json.load(config_file)
-
-    if isinstance(param, str):
-        try:
-            return config[param]
-        except KeyError:
-            return "KeyError"
-
-    else:
-        items = []
-        for item in param:
-            if item in config:
-                items.append(config[item])
-            else:
-                items.append(None)
-        return items
-
-
-def var_set(param, value):
-    """
-        Allow to edit config parameters
-    """
-    with open("json/config.json", "r", encoding="utf-8") as config_file:
-        config = json.load(config_file)
-
-    if isinstance(param, str):
-        config[param] = value
-
-    elif isinstance(param, list):
-        for i in range(len(param)):
-            config[param[i]] = value[i]
-
-    else:
-        raise Exception("Invalid param type")
-
-    with open("json/config.json", "w", encoding="utf-8") as config_file:
-        json.dump(config, config_file, indent=2)
-
-
-def add_event(timestamp: int, data: dict):
-    with open("json/events.json", "r", encoding="utf-8") as event_file:
-        events = json.load(event_file)
-
-    if timestamp in events:
-        events[timestamp].append(data)
-    else:
-        events[timestamp] = [data]
-
-    with open("json/config.json", "w", encoding="utf-8") as event_file:
-        json.dump(events, event_file, indent=2)
-
-
-def get_text(reference: str, lang: str):
-    """
-        Used to get the corresponding text in language asked
-    """
-    # make it open the corresponding file and return the associated text (return reference if error)
-    #   if lang == "", open english file
-    return reference
-
-
-def key_with_lowest_value(dico: dict):
-    min_value = None
-    min_key = None
-    for item in dico.items():
-        try:
-            if item[1] < min_value:
-                min_value = item[1]
-                min_key = item[0]
-
-        except TypeError:
-            min_value = item[1]
-            min_key = item[0]
-    return min_key
-
-
-def sort_dict_by_value(d, reverse=False):
-    return dict(sorted(d.items(), key=lambda x: x[1], reverse=reverse))
-
-
-def string_to_time(raw_time: str):
-    time_dict = {
-        "s": 1,  # second
-        "m": 60,  # minute
-        "h": 3600,  # hour
-        "d": 3600 * 24,  # day
-        "w": 3600 * 24 * 7,  # week
-        "y": 3600 * 24 * 7 * 28 * 12,  # year
-    }
-    times = raw_time.split(' ')
-    total = 0
-
-    for time in times:
-        if time == '':
-            continue
-        indicator = time[-1]
-
-        if indicator not in time_dict:
-            raise InvalidTimeString(reason="Invalid indicator", raw_input=raw_time)
-
-        try:
-            number = eval(time[:-1])
-
-        except SyntaxError:
-            raise InvalidTimeString(reason="Invalid time", raw_input=raw_time)
-
-        if number < 0:
-            raise InvalidTimeString(reason="Negative time isn't allowed", raw_input=raw_time)
-
-        total += number * time_dict[indicator]
-
-    return total
-
-
-def time_to_string(raw_time: int):
-    result = ""
-
-    if raw_time >= 3600 * 24 * 365:
-        year_amt = raw_time // (3600 * 24 * 365)
-        raw_time -= year_amt * (3600 * 24 * 365)
-        result += f"{year_amt}y "
-
-    if raw_time >= 3600 * 24 * 7:
-        week_amt = raw_time // (3600 * 24 * 7)
-        raw_time -= week_amt * (3600 * 24 * 7)
-        result += f"{week_amt}w "
-
-    if raw_time >= 3600 * 24:
-        day_amt = raw_time // (3600 * 24)
-        raw_time -= day_amt * (3600 * 24)
-        result += f"{day_amt}d "
-
-    if raw_time >= 3600:
-        hour_amt = raw_time // 3600
-        raw_time -= hour_amt * 3600
-        result += f"{hour_amt}h "
-
-    if raw_time > 60:
-        minute_amt = raw_time // 60
-        raw_time -= minute_amt * 60
-        result += f"{minute_amt}m "
-
-    if raw_time > 1:
-        second_amt = raw_time
-        raw_time -= second_amt
-        result += f"{second_amt}s "
-
-    return result[:-1]
 
 
 class MyBot(commands.Bot):
@@ -403,6 +240,12 @@ class MyBot(commands.Bot):
             # emb = Embed(color=Color.fuchsia(),  description=get_text("command.disabled", user_lang))
             emb = Embed(color=Color.fuchsia(), description=f"Command disabled ({exception.reason})")
 
+        elif isinstance(exception, UnknownCode):
+            emb = Embed(color=Color.red(),  description=get_text("error.unknown_code", user_lang))
+
+        elif isinstance(exception, CodeLimitReached):
+            emb = Embed(color=Color.red(),  description=get_text("error.code_limit_reached", user_lang))
+
         else:
             self.log_action(f"Unhandled error occurred ({type(exception)}) : {exception}", self.cmd_logger, 50)
             adm_emb = Embed(color=Color.red(), description=f"Error `{type(exception)}` cannot be handled. "
@@ -446,15 +289,34 @@ async def event_loop():
     for timestamp in events:
         if timestamp <= now:
             for item in events[timestamp]:
-                guild = bot.get_guild(item['guild'])
-                user = guild.get_member(item['user'])
 
                 if item['type'] == 'tempmute':
+                    guild = bot.get_guild(item['guild'])
+                    user = guild.get_member(item['user'])
                     mute_role = guild.get_role(get_parameter('mute_role'))
+
                     await user.remove_roles(mute_role, reason="End of sentence")
+                    bot.log_action(f"[MOD] {user} has been un-muted (automatically)", bot.mod_logger)
 
                 elif item['type'] == 'tempban':
+                    guild = bot.get_guild(item['guild'])
+                    user = guild.get_member(item['user'])
+
                     await user.unban(reason="End of sentence")
+                    bot.log_action(f"[MOD] {user} has been un-banned (automatically)", bot.mod_logger)
+
+                elif item['type'] == "coins-boost":
+                    var_set("global_coins_boost", item['value'])
+                    bot.log_action(f"[ECO] Global boost value is now {item['value']}", bot.eco_logger)
+
+                else:
+                    bot.log_action(f"[LOOP] Unknown event type : '{item['type']}'", bot.bot_logger, 30)
+                    continue
+
+                del item
+
+    with open("json/events.json", "w", encoding="utf-8") as events_file:
+        json.dump(events, events_file, indent=2)
 
 
 @event_loop.before_loop
